showing customer profile (21)
the site now shows a customer profile when clicking the corresponding 'show profile' button. currently I get a single error when I load up the page as my code attempts to load data from an address that doesn't exist (due to the ID not being selected) - this doesn't cause any functional issue but I'd like to know how to fix it.