3. Created and used a new component to show info cards
In src/App.js, above the <Bookings /> component, added a new component called TouristInfoCards which shows 3 cards. A card is a common user interface pattern with an image at the top and some related text underneath. The cards links to peoplemakeglasgow.com, visitmanchester.com and visitlondon.com. The cards contains the name of the city and an image of the city.