4. Create a Footer component
Created a <Footer /> component which is rendered at the bottom of the page. Passed the following array as a prop to this component: ["123 Fake Street, London, E1 4UD", "hello@fakehotel.com", "0123 456789"]. Inside the component, used the data we passed as a prop to render a <ul> list with each item of the array displayed as a <li>.

Tested: The footer rendered at the bottom of the page with each address property displayed as a list item.