6. Showed more bookings in the table
Instead of using hard-coded data in the <SearchResults /> component, I loaded data from the src/data/fakeBookings.json file in the <Bookings /> component and passed it as a prop to <SearchResults />. All the bookings in src/data/fakeBookings.json are now be displayed in my table.

Hint: Look in the <Bookings /> component for how to import data from a JSON file.

Tested: All the bookings in the file src/data/fakeBookings.json are displayed in my table.