7. Calculated and showed the number of nights for each booking
Added another column to my <SearchResults /> table which shows the number of nights a guest is staying.

Tried installing the moment.js library (by installing it with npm install moment --save) and using the .diff() method to compare dates.

Tested: Each booking in my table and showed the number of nights in a separate column. For example, Mr John Doe has a booking for 2 nights.