9. Prepared to add more pizzas
Previously, the number of pizzas a guest can order was static and set to 0, even if they click on the 'Add' button. So I changed that in the following to let a guest add more pizzas to their order.
First, I declared a new state variable orders along with the function to set the orders state setOrders. The initial value of the orders state should be 0. I used the new orders variable instead of the pizzas variable (that I can now delete).

I used the React function useState to create a state variable. I imported the function at the top with import React, {useState} from "react";.

Verified the number of ordered pizzas it still 0 on the screen and did a bit of styling with CSS to make it look better.