10. Adding more pizzas
In the <Restaurant /> component, I created a new function named orderOne. The orderOne function doesn't take any parameters. I used the setOrders function to increment the orders state variable by 1. Then, added an onClick handler to the Add <button> that calls the orderOne function when it's being clicked.

Tested: Click on the Add button a few times and verify that the number of pizzas increases accordingly and it did. :)