11. Extracted the Add button to its own component
(Corrected the name of the previous function we created to orderOne from odederOne.)
Extracted the <button> currently in the <Restaurant /> component to a new component named RestaurantButton. Passed the orderOne function as a prop to the <RestaurantButton /> component and used this prop in the onClick handler.

Tested: Clicking the button still increments the number of pizzas.