12. Extracted pizza order to its own Order component
Extracted the <li> containing "Pizzas" from the <Restaurant /> component to a new component named Order. Also, moved the declaration of the orders state and the orderOne function from the <Restaurant /> component to the new <Order /> component. Used the <Order /> component in the <ul> list of the <Restaurant /> component.

Tested: Made sure the pizza order is still rendered on the page and that clicking on the "Add" button still increments the number of orders.