13. Rendered more orders
Paseds a new prop named orderType to the <Order /> component with the value "Pizzas". Then rendered the orderType prop instead of "Pizzas" in the <Order /> component. Made sure that "Pizzas" is still displayed on the screen. In the <ul> list of the <Restaurant /> component, rendered 2 others <Order /> components but this time pass different values for the orderType prop: "Salads" and "Chocolate cake".

Test: For each order, the number of items are incremented independently. I am able to explain what is happening. ;)