14. Passing bookings from a state variable
In the <Bookings /> component, declared a new state bookings with the corresponding setter function setBookings to hold the FakeBookings data. Instead of passing FakeBookings directly to the <SearchResults /> component, passed the new bookings state variable.

The new bookings state is  initialised with the FakeBookings variable.

Checked that the bookings are still rendered correctly in the page.