15. a) Highlighted single booking row when clicked
Within the <SearchResults /> component or its child components, added an onClick handler to each row in the table (on the <tr> element). When clicked, the row is "selected" and highlighted with a different colour. When clicked again, the row is unselected and the coloured highlighting is removed.

We need to get more than one rows highlighted.