16. Load bookings remotely
Instead of getting the existing bookings from the file data/fakeBookings.json, we got and load the bookings from a remote API. In the <Bookings /> component, we used the React function useEffect to console.log() some text only when the page first renders on the screen. We verified that when we refreshed the page, the text appears once in the console. Then, in the useEffect function, use the fetch() function to get data from https://cyf-react.glitch.me.

Hints:
- Replaced FakeBookings in the bookings state and initialised it with [] (because we haven't fetched any results yet!)
- After calling the fetch() function, we used .then() to handle the response. Tried looking at your Pokemon app that you worked on in class for an example
- When the response came back, use setBookings to update the results

Test: Verified the customers data are still displayed correctly in the table.