17. Storing the search input in a state
We have implemented the functionality to search for a customer name given the text typed into the customer name field. In the src/Search.js file, we declared a new state variable named searchInput with the corresponding setter function setSearchInput (hint: use the React function useState). The initial value of the searchInput variable is an empty string. We added a value property to the <input> tag that is set to the new searchInput state variable. Then we created a new function handleSearchInput taking an event parameter. This function used the setSearchInput function to update the state variable searchInput with what the user typed in the input field. Finally, added an onChange prop to the <input> tag that is set to the function handleSearchInput. Used console.log() to output the value received in the handleSearchInput function.

Hint: We used event.target.value to get the input value.

Test: In the developer console, checked that everything we type in the search input is printed successively for each new character we entered.