18. Triggering search when submitting the form
Still in the <Search /> component, we added an onSubmit handler to the <form> tag. When the form is submitted (tried clicking the search button), got the value of the state searchInput and passed it as a parameter to the search prop function that had been provided for us (the search prop is passed from the <Bookings /> component).

Note: Also our submit handler should take an event parameter and add the line event.preventDefault() to prevent the browser to implicitly submit the form).

Test: Looked in the console, we should see the text that is typed in the search input field when submitting the form.