19. Implementing the search functionality and refactoring the SearchResults.jsx
Still in the <Bookings /> component, we implement the search method. It must use the searchVal variable (that we just passed from the <Search /> component) to filter the search results. The filter function should return bookings where firstName or surname match searchVal. Once filtered, we use the setBookings function to update the results rendered in <SearchResults />.

Test: Verified that when you enter an existing first name or surname and submit the form, the results are filtered accordingly in the customers table.