22. Show a loading message
We remember in Lesson 2, we fetched the bookings from a remote API. Now we are showing a loading state in <Bookings /> while the data from the server is being fetched. To test this, we try loading data from https://cyf-react.glitch.me/delayed, which has a 5 second delay before returning the data. We will need to use another state to record when our application is loading data (this can be a boolean) and display a loading message whenever the application is loading data.

Hint: We try looking at our Pokemon app that we worked on as an example.

Tested: A message inviting the user to wait should be displayed on the screen until bookings data is rendered on the screen. When bookings are rendered, the loading message should be hidden.